Auth fix
