--- src/routers/auth/authLogin.js
@@ -1,19 +1,0 @@
-import express from 'express';
-
-import { validateBody } from '../../middlewares/validateBody.js';
-import { userLoginSchema } from '../../validation/user.js';
-import { ctrlWrapper } from '../../utils/ctrlWrapper.js';
-import { loginController } from '../../controller/auth/controllersLogin.js';
-
-const router = express.Router();
-
-const jsonParser = express.json();
-
-router.post(
-  '/login',
-  jsonParser,
-  validateBody(userLoginSchema),
-  ctrlWrapper(loginController),
-);
-
-export default router;
