--- src/routers/auth/authRegister.js
@@ -1,19 +1,0 @@
-import express from 'express';
-
-import { validateBody } from '../../middlewares/validateBody.js';
-import { userRegisterSchema } from '../../validation/user.js';
-import { ctrlWrapper } from '../../utils/ctrlWrapper.js';
-import { registerController } from '../../controller/auth/controllersRegister.js';
-
-const router = express.Router();
-
-const jsonParser = express.json();
-
-router.post(
-  '/register',
-  jsonParser,
-  validateBody(userRegisterSchema),
-  ctrlWrapper(registerController),
-);
-
-export default router;
